chore: Move types file into "lib" direcotry
